Move code to module specific dir
