+++ services/blockchain-indexer/shared/dataService/business/index.js
@@ -71,6 +71,10 @@ const {
 } = require('./commandsParamsSchemas');
 
 const {
+	getVotes,
+} = require('./dpos/votesSent');
+
+const {
 	getVotesByTransactionIDs,
 } = require('./voters');
 
@@ -108,4 +112,5 @@ module.exports = {
 	getCommandsParamsSchemas,
 	normalizeTransaction,
 	getVotesByTransactionIDs,
+	getVotes,
 };
